requirements.txt
streamlit
openai
yfinance
plotly
python-dotenv